Split menu logic from object file
