+++ src/object/C4Object.cpp
@@ -24,15 +24,12 @@
 #include "control/C4Record.h"
 #include "game/C4Physics.h"
 #include "game/C4Viewport.h"
-#include "graphics/C4Draw.h"
-#include "graphics/C4GraphicsResource.h"
 #include "gui/C4GameMessage.h"
 #include "landscape/C4MaterialList.h"
 #include "landscape/C4PXS.h"
 #include "landscape/C4Particles.h"
 #include "landscape/C4SolidMask.h"
 #include "landscape/fow/C4FoW.h"
-#include "lib/StdColors.h" // This was imported via C4GameMessage.h, but should be listed extra imo
 #include "object/C4Command.h"
 #include "object/C4Def.h"
 #include "object/C4DefList.h"
@@ -48,6 +45,7 @@
 #include "player/C4RankSystem.h"
 #include "script/C4AulExec.h"
 #include "script/C4Effect.h"
+
 
 void C4Action::SetBridgeData(int32_t iBridgeTime, bool fMoveClonk, bool fWall, int32_t iBridgeMaterial)
 {
@@ -928,132 +926,6 @@ bool C4Object::Lift(C4Real tydir, C4Real dforce)
 	return true;
 }
 
-static void DrawMenuSymbol(int32_t iMenu, C4Facet &cgo, int32_t iOwner)
-{
-	C4Facet ccgo;
-
-	DWORD dwColor=0;
-	if (ValidPlr(iOwner)) dwColor=::Players.Get(iOwner)->ColorDw;
-
-	switch (iMenu)
-	{
-	case C4MN_Buy:
-		::GraphicsResource.fctFlagClr.DrawClr(ccgo = cgo.GetFraction(75, 75), true, dwColor);
-		::GraphicsResource.fctWealth.Draw(ccgo = cgo.GetFraction(100, 50, C4FCT_Left, C4FCT_Bottom));
-		::GraphicsResource.fctArrow.Draw(ccgo = cgo.GetFraction(70, 70, C4FCT_Right, C4FCT_Center), false, 0);
-		break;
-	case C4MN_Sell:
-		::GraphicsResource.fctFlagClr.DrawClr(ccgo = cgo.GetFraction(75, 75), true, dwColor);
-		::GraphicsResource.fctWealth.Draw(ccgo = cgo.GetFraction(100, 50, C4FCT_Left, C4FCT_Bottom));
-		::GraphicsResource.fctArrow.Draw(ccgo = cgo.GetFraction(70, 70, C4FCT_Right, C4FCT_Center), false, 1);
-		break;
-	}
-}
-
-bool C4Object::ActivateMenu(int32_t iMenu, int32_t iMenuSelect,
-                            int32_t iMenuData, int32_t iMenuPosition,
-                            C4Object *pTarget)
-{
-	// Variables
-	C4FacetSurface fctSymbol;
-	C4IDList ListItems;
-	// Close any other menu
-	if (Menu && Menu->IsActive()) if (!Menu->TryClose(true, false)) return false;
-	// Create menu
-	if (!Menu) Menu = new C4ObjectMenu; else Menu->ClearItems();
-	// Open menu
-	switch (iMenu)
-	{
-		// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	case C4MN_Activate:
-		// No target specified: use own container as target
-		if (!pTarget) if (!(pTarget=Contained)) break;
-		// Opening contents menu blocked by RejectContents
-		if (!!pTarget->Call(PSF_RejectContents)) return false;
-		// Create symbol
-		fctSymbol.Create(C4SymbolSize,C4SymbolSize);
-		pTarget->Def->Draw(fctSymbol,false,pTarget->Color,pTarget);
-		// Init
-		Menu->Init(fctSymbol,FormatString(LoadResStr("IDS_OBJ_EMPTY"),pTarget->GetName()).getData(),this,C4MN_Extra_None,0,iMenu);
-		Menu->SetPermanent(true);
-		Menu->SetRefillObject(pTarget);
-		// Success
-		return true;
-		// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	case C4MN_Buy:
-		// No target specified: container is base
-		if (!pTarget) if (!(pTarget=Contained)) break;
-		// Create symbol
-		fctSymbol.Create(C4SymbolSize,C4SymbolSize);
-		DrawMenuSymbol(C4MN_Buy, fctSymbol, pTarget->Owner);
-		// Init menu
-		Menu->Init(fctSymbol,LoadResStr("IDS_PLR_NOBUY"),this,C4MN_Extra_Value,0,iMenu);
-		Menu->SetPermanent(true);
-		Menu->SetRefillObject(pTarget);
-		// Success
-		return true;
-		// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	case C4MN_Sell:
-		// No target specified: container is base
-		if (!pTarget) if (!(pTarget=Contained)) break;
-		// Create symbol & init
-		fctSymbol.Create(C4SymbolSize,C4SymbolSize);
-		DrawMenuSymbol(C4MN_Sell, fctSymbol, pTarget->Owner);
-		Menu->Init(fctSymbol,FormatString(LoadResStr("IDS_OBJ_EMPTY"),pTarget->GetName()).getData(),this,C4MN_Extra_Value,0,iMenu);
-		Menu->SetPermanent(true);
-		Menu->SetRefillObject(pTarget);
-		// Success
-		return true;
-		// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	case C4MN_Get:
-	case C4MN_Contents:
-		// No target specified
-		if (!pTarget) break;
-		// Opening contents menu blocked by RejectContents
-		if (!!pTarget->Call(PSF_RejectContents)) return false;
-		// Create symbol & init
-		fctSymbol.Create(C4SymbolSize,C4SymbolSize);
-		pTarget->Def->Draw(fctSymbol,false,pTarget->Color,pTarget);
-		Menu->Init(fctSymbol,FormatString(LoadResStr("IDS_OBJ_EMPTY"),pTarget->GetName()).getData(),this,C4MN_Extra_None,0,iMenu);
-		Menu->SetPermanent(true);
-		Menu->SetRefillObject(pTarget);
-		// Success
-		return true;
-		// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	case C4MN_Info:
-		// Target by parameter
-		if (!pTarget) break;
-		// Create symbol & init menu
-		fctSymbol.Create(C4SymbolSize, C4SymbolSize); GfxR->fctOKCancel.Draw(fctSymbol,true,0,1);
-		Menu->Init(fctSymbol, pTarget->GetName(), this, C4MN_Extra_None, 0, iMenu, C4MN_Style_Info);
-		Menu->SetPermanent(true);
-		Menu->SetAlignment(C4MN_Align_Free);
-		C4Viewport *pViewport = ::Viewports.GetViewport(Controller); // Hackhackhack!!!
-		if (pViewport) Menu->SetLocation((pTarget->GetX() + pTarget->Shape.GetX() + pTarget->Shape.Wdt + 10 - pViewport->GetViewX()) * pViewport->GetZoom(),
-			                                 (pTarget->GetY() + pTarget->Shape.GetY() - pViewport->GetViewY()) * pViewport->GetZoom());
-		// Add info item
-		fctSymbol.Create(C4PictureSize, C4PictureSize); pTarget->Def->Draw(fctSymbol, false, pTarget->Color, pTarget);
-		Menu->Add(pTarget->GetName(), fctSymbol, "", C4MN_Item_NoCount, nullptr, pTarget->GetInfoString().getData());
-		fctSymbol.Default();
-		// Success
-		return true;
-		// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	}
-	// Invalid menu identification
-	CloseMenu(true);
-	return false;
-}
-
-bool C4Object::CloseMenu(bool fForce)
-{
-	if (Menu)
-	{
-		if (Menu->IsActive()) if (!Menu->TryClose(fForce, false)) return false;
-		if (!Menu->IsCloseQuerying()) { delete Menu; Menu=nullptr; } // protect menu deletion from recursive menu operation calls
-	}
-	return true;
-}
-
 BYTE C4Object::GetArea(int32_t &aX, int32_t &aY, int32_t &aWdt, int32_t &aHgt) const
 {
 	if (!Status || !Def) return 0;
@@ -1503,13 +1375,6 @@ void C4Object::Clear()
 	delete pDrawTransform; pDrawTransform = nullptr;
 	delete pGfxOverlay; pGfxOverlay = nullptr;
 	delete pMeshInstance; pMeshInstance = nullptr;
-}
-
-bool C4Object::MenuCommand(const char *szCommand)
-{
-	// Native script execution
-	if (!Def || !Status) return false;
-	return !! ::AulExec.DirectExec(this, szCommand, "MenuCommand");
 }
 
 void C4Object::SetSolidMask(int32_t iX, int32_t iY, int32_t iWdt, int32_t iHgt, int32_t iTX, int32_t iTY)
